docs(relay-client): add some doc comments
